Ensure output port is brought low to make reads work.
When disabling driving of the data lines, we also need to bring the
output lines low in order to not continue to pull them high, so that
we can read effectively.

I also added an improved test routine to write and read more
addresses.

It still doesn't produce the output expected, but it's a lot closer!