Add a delay to the SIMM read - now works!
Turns out we were reading the data "too early" because the dual
flip-flop synchronizer meant we were seeing the data from before CAS
was asserted. One NOP, and the data lines up!

Sped up the clock to 16MHz, still functions nicely.

Also added a couple of minor tweaks to the main loop to run it a
little more quickly.