Add the start of a results-parser, tidy up results files
Huh, the results files contain a lot more corruption than I was
expecting. In an ideal world I'd be digging around to find where the
bytes are being dropped (in many cases, in quite a predictable way),
and I'd be dropping corrupted entries. In the common case, though,
it's clear what the output would be (no bits dropped at millisecond
intervals), so I just made up those results. This would make me an
evil scientist if any of this mattered, but it doesn't really. ;)